Section 4, Problem Solved for AI Dance or Under terrain
First, need to thanks Richard to provide me this solution and solved it

Solution: Fixed after finding a solution on Ben's game forums, thanks user Maxximum for the idea!  Disable 'Simulate Physics' on the Tank root.  Then in Tank_PB Event Graph, set 'Simulate Physics' back to true as part of Event BeginPlay.  A bit of a roundabout fix, but it now works.